consolidate sysinfo and tokio-metrics examples into extras
